Reinsert unfinished DAG runInfos in the main BG_WORKER loop. Refactor tests - separate tests by using classes
